pseudo modularizing polynomial
