+++ include/eve/module/polynomial/detail/newton_impl.hpp
@@ -7,19 +7,8 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/traits/common_compatible.hpp>
-#include <eve/concept/compatible.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/detail/apply_over.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/decorator.hpp>
-#include <eve/function/fma.hpp>
-#include <eve/function/regular.hpp>
-#include <eve/function/sub.hpp>
-#include <eve/constant/one.hpp>
+#include <eve/module/arithmetic.hpp>
 #include <eve/concept/range.hpp>
-
 #include <concepts>
 #include <iterator>
 #include <initializer_list>
