+++ include/eve/module/polynomial/diff/impl/horner.hpp
@@ -7,14 +7,7 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/compatible.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/constant/zero.hpp>
-#include <eve/detail/apply_over.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/derivative.hpp>
-#include <eve/function/pedantic/fma.hpp>
+#include <eve/module/arithmetic.hpp>
 
 namespace eve::detail
 {
