+++ include/eve/module/polynomial/diff/impl/reverse_horner.hpp
@@ -7,13 +7,9 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/compatible.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/constant/zero.hpp>
-#include <eve/detail/apply_over.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/detail/skeleton_calls.hpp>
-#include <eve/function/derivative.hpp>
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/polynomial/regular/reverse_horner.hpp>
+#include <eve/module/polynomial/pedantic/reverse_horner.hpp>   
 
 namespace eve::detail
 {
