+++ include/eve/module/polynomial/numeric/impl/reverse_horner.hpp
@@ -7,10 +7,6 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/compatible.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/constant/one.hpp>
-#include <eve/function/numeric/fma.hpp>
 #include <eve/module/polynomial/detail/reverse_horner_impl.hpp>
 
 namespace eve::detail
