+++ include/eve/module/polynomial/regular/impl/reverse_horner.hpp
@@ -7,10 +7,6 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/compatible.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/function/regular.hpp>
-#include <eve/constant/one.hpp>
 #include <eve/module/polynomial/detail/reverse_horner_impl.hpp>
 
 namespace eve::detail
