+++ include/eve/module/polynomial/regular/kind.hpp
@@ -14,6 +14,34 @@ namespace eve
   //================================================================================================
   //================================================================================================
   // Function decorators mark-up used in function overloads
+  struct p_kind_
+  {
+    template<typename D> static constexpr auto combine( D const& ) noexcept =delete;
+  };
+
+  using p_kind_type = decorated<p_kind_()>;
+  //================================================================================================
+  //! @addtogroup decorator
+  //! @{
+  //! @var p_kind
+  //!
+  //! @brief  Higher-order @callable imbuing p_kind behaviour onto other @callable{s}.
+  //!
+  //! #### Synopsis
+  //!
+  //! @param f
+  //! An instance of eve::callable
+  //!
+  //! @return
+  //! A @callable performing the p_kind algorithm choice
+  //!
+  //!  @}
+  //================================================================================================
+  inline constexpr p_kind_type const p_kind = {};
+
+  //================================================================================================
+  //================================================================================================
+  // Function decorators mark-up used in function overloads
   struct q_kind_
   {
     template<typename D> static constexpr auto combine( D const& ) noexcept =delete;
