+++ include/eve/module/polynomial/regular/polynomial.hpp
@@ -10,7 +10,9 @@
 #include <eve/module/polynomial/regular/gegenbauer.hpp>
 #include <eve/module/polynomial/regular/hermite.hpp>
 #include <eve/module/polynomial/regular/horner.hpp>
+#include <eve/module/polynomial/regular/kind.hpp>
 #include <eve/module/polynomial/regular/laguerre.hpp>
 #include <eve/module/polynomial/regular/legendre.hpp>
 #include <eve/module/polynomial/regular/newton.hpp>
 #include <eve/module/polynomial/regular/reverse_horner.hpp>
+#include <eve/module/polynomial/regular/successor.hpp>
